HBASE-27448 Add an admin method to get replication enabled state
