+++ hbase-client/src/main/java/org/apache/hadoop/hbase/client/AsyncAdmin.java
@@ -796,6 +796,14 @@ public interface AsyncAdmin {
   CompletableFuture<Void> disableTableReplication(TableName tableName);
 
   /**
+   * Check if a replication peer is enabled.
+   * @param peerId id of replication peer to check
+   * @return true if replication peer is enabled. The return value will be wrapped by a
+   *         {@link CompletableFuture}.
+   */
+  CompletableFuture<Boolean> isReplicationPeerEnabled(String peerId);
+
+  /**
    * Take a snapshot for the given table. If the table is enabled, a FLUSH-type snapshot will be
    * taken. If the table is disabled, an offline snapshot is taken. Snapshots are taken sequentially
    * even when requested concurrently, across all tables. Snapshots are considered unique based on
@@ -1776,12 +1784,4 @@ public interface AsyncAdmin {
    * Flush master local region
    */
   CompletableFuture<Void> flushMasterStore();
-
-  /**
-   * Check if a replication peer is enabled.
-   * @param peerId id of replication peer to check
-   * @return true if replication peer is enabled. The return value will be wrapped by a
-   *         {@link CompletableFuture}.
-   */
-  CompletableFuture<Boolean> isReplicationPeerEnabled(String peerId);
 }
