+++ hbase-server/src/test/java/org/apache/hadoop/hbase/rsgroup/VerifyingRSGroupAdmin.java
@@ -676,6 +676,11 @@ public class VerifyingRSGroupAdmin implements Admin, Closeable {
     return admin.transitReplicationPeerSyncReplicationStateAsync(peerId, state);
   }
 
+  @Override
+  public boolean isReplicationPeerEnabled(String peerId) throws IOException {
+    return admin.isReplicationPeerEnabled(peerId);
+  }
+
   public void decommissionRegionServers(List<ServerName> servers, boolean offload)
     throws IOException {
     admin.decommissionRegionServers(servers, offload);
@@ -938,10 +943,5 @@ public class VerifyingRSGroupAdmin implements Admin, Closeable {
   @Override
   public void flushMasterStore() throws IOException {
     admin.flushMasterStore();
-  }
-
-  @Override
-  public boolean isReplicationPeerEnabled(String peerId) throws IOException {
-    return admin.isReplicationPeerEnabled(peerId);
   }
 }
