+++ hbase-thrift/src/main/java/org/apache/hadoop/hbase/thrift2/client/ThriftAdmin.java
@@ -1032,6 +1032,11 @@ public class ThriftAdmin implements Admin {
   }
 
   @Override
+  public boolean isReplicationPeerEnabled(String peerId) throws IOException {
+    throw new NotImplementedException("isReplicationPeerEnabled not supported in ThriftAdmin");
+  }
+
+  @Override
   public void decommissionRegionServers(List<ServerName> servers, boolean offload) {
     throw new NotImplementedException("decommissionRegionServers not supported in ThriftAdmin");
 
@@ -1313,10 +1318,5 @@ public class ThriftAdmin implements Admin {
   @Override
   public void flushMasterStore() throws IOException {
     throw new NotImplementedException("flushMasterStore not supported in ThriftAdmin");
-  }
-
-  @Override
-  public boolean isReplicationPeerEnabled(String peerId) throws IOException {
-    throw new NotImplementedException("isReplicationPeerEnabled not supported in ThriftAdmin");
   }
 }
